Adds USSD balance endpoint for dashboard.

+++ pkg/service/client/NotifyApi.go
@@ -12,7 +12,7 @@ type NotifyApiClient struct {
 }
 type USSDBalanceApiResponse struct {
 	ApiResponse
-	Data *int `json:"data"`
+	Data *float64 `json:"data"`
 }
 
 func InitNotifyClient() *NotifyApiClient {
@@ -33,7 +33,7 @@ func (n *NotifyApiClient) SendNotification(request *NotificationRequest) error {
 	return nil
 }
 
-func (n *NotifyApiClient) GetUSSDBalance() (int, error) {
+func (n *NotifyApiClient) GetUSSDBalance() (float64, error) {
 	res := new(USSDBalanceApiResponse)
 
 	if err := n.newRequest(http.MethodGet, "/dashboard/ussd-balance", nil).send(&res); err != nil {
