+++ pkg/service/notify.go
@@ -1,6 +1,6 @@
 package service
 
-func GetUSSDBalance() (int, error) {
+func GetUSSDBalance() (float64, error) {
 	balance, err := notifyClient.GetUSSDBalance()
 	if err != nil {
 		return 0, err
